Renaming AbTestService to AbExperiment
I'm loath to end any production class's filename with "test".  I don't
want minitest or rspec coming along and attempting to "test" that file.
In addition, I find that Experiment better describes the concept.

Also, per conversations with @citizen428, we favor nouns objects to go
in app/models and verb objects to go in app/services.  That's a great
separation.